+++ lightly_studio_view/src/lib/hooks/index.ts
@@ -18,3 +18,5 @@ export { useVideo } from '$lib/hooks/useVideo/useVideo';
 export { useAddTagToSample } from '$lib/hooks/useAddTagToSample/useAddTagToSample';
 export { useFileDrop } from '$lib/hooks/useFileDrop/useFileDrop';
 export { useImageUpload } from '$lib/hooks/useImageUpload/useImageUpload';
+export { useFeatureFlags } from '$lib/hooks/useFeatureFlags/useFeatureFlags';
+export { useSelectAll } from '$lib/hooks/useSelectAll/useSelectAll';